+++ src/game/client/c_baseanimatingoverlay.cpp
@@ -112,7 +112,7 @@ void C_AnimationLayer::SetWeight( float flWeight )
 	m_flWeight = flWeight;
 }
 
-C_BaseAnimatingOverlay::C_BaseAnimatingOverlay( int iEFlags )
+C_BaseAnimatingOverlay::C_BaseAnimatingOverlay( uint64 iEFlags )
 	: C_BaseAnimating( iEFlags )
 {
 	// FIXME: where does this initialization go now?
